Add a test widget, pull notif content from web
Added a (broken) widget to the build.
Added a new Kotlin source file (MakeNotif) which is a AsyncTask. It pulls down humans.txt and uses that for the source of the notification text. If there is an error, it is shown to the user via a toast.

Added the internet permission
Added OkHttp3 for accessing external stuff